addition of subfolder for the results
